requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==9.5.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu
requests>=2.28.0
numpy>=1.21.0,<2.0.0
--extra-index-url https://download.pytorch.org/whl/cpu